requirements.txt
python-telegram-bot==20.7
python-dotenv==1.1.1
APScheduler==3.6.3
requests==2.32.3